Refactor #30 - 비밀번호 인증 기능 위치 변경(auth controller -> member controller)
