+++ src/main/java/com/maruhxn/boardserver/service/MemberService.java
@@ -3,6 +3,7 @@ package com.maruhxn.boardserver.service;
 import com.maruhxn.boardserver.common.exception.ErrorCode;
 import com.maruhxn.boardserver.common.exception.GlobalException;
 import com.maruhxn.boardserver.domain.Member;
+import com.maruhxn.boardserver.dto.request.auth.ConfirmPasswordRequest;
 import com.maruhxn.boardserver.dto.request.members.UpdateMemberProfileRequest;
 import com.maruhxn.boardserver.dto.request.members.UpdatePasswordRequest;
 import com.maruhxn.boardserver.dto.response.object.MemberItem;
@@ -37,6 +38,20 @@ public class MemberService {
                 .email(findMember.getEmail())
                 .profileImage(findMember.getProfileImage())
                 .build();
+    }
+
+    /**
+     * 비밀번호 확인(인증)
+     *
+     * @param memberId
+     * @param confirmPasswordRequest
+     */
+    public void confirmPassword(Long memberId, ConfirmPasswordRequest confirmPasswordRequest) {
+        Member findMember = memberRepository.findById(memberId)
+                .orElseThrow(() -> new GlobalException(ErrorCode.NOT_FOUND_USER));
+        Boolean isMatch = passwordEncoder.matches(confirmPasswordRequest.getCurrPassword(), findMember.getPassword());
+
+        if (!isMatch) throw new GlobalException(ErrorCode.INCORRECT_PASSWORD);
     }
 
     /**
